requirements.txt
flask
Flask-SQLAlchemy
mysql-connector-python
python-dotenv